Added support for running the PnP_ROM_Compile.tcl script directly from the Quartus flow. This script is requiring the tDOM XML parser library. This library was compiled for a TCL shell running inside Quartus  64 bits. See
backend/tdom/win64/Readme.txt

for more info on how to re-compile the library if required.

I also added a build ID number on the QUARTUS compilation revision. This build ID is generated from the Unix Epoch (in seconds). Note that the Same build ID could be stored in the PnP_ROM_Compile ROM file. This is not implemented yet.